api: fix high port numbers
and convert ccminer.rc from utf-16 to Latin1 (doesnt want utf-8),
required for git diff/merge/rebase...

+++ api.cpp
@@ -722,7 +722,7 @@ static bool check_connect(struct sockaddr_in *cli, char **connectaddr, char *gro
 static void api()
 {
 	const char *addr = opt_api_allow;
-	short int port = opt_api_listen; // 4068
+	unsigned short port = (unsigned short) opt_api_listen; // 4068
 	char buf[MYBUFSIZ];
 	int c, n, bound;
 	char *connectaddr;
